font-patcher: Handle Powerline glyphs always as monospaced
[why]
The Powerline glyphs can have strange width if a non-mono font is
created. The width of the glyphs is then based on the symbol font which
is not always useful. Some get a very slim advance width.

[how]
Temporarely turn on '--mono' handling of the glyphs if a Powerline
codepoint is processed.

Fixes: #1695

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>