font-patcher: Keep Powerline glyphs in one cell
[why]
When we have a 'stubby' font, i.e. one with a square-ish cell the 'pa'
scaled Powerline icons are just too big.

Normally 'stubby' fonts prevent scaling into 2 cells - with 'xy'
scaling.

For the Powerline glyphs - which are fitted into the cells - we want to
avoid scaling so big and the Powerline glyphs should keep a square or
slimmer appearance.

[how]
Enforce the '1' scaling rule for powerline glyphs in 'stubby' fonts.

Relocate some code to have this in the right spot before anything is
done with the scale options.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>